Level Select Menu's interface updated, Garage updated,
I updated the level select menu's so each level changes to a green icon once completed in both the city menu and the mine menu,
I updated the garage to disable the select button on certain vehicles that need to be unlocked. The BMW and SUV unlocked in the City and Mine bonus stages respectively. The other remaining vehicles will be locked until
completion of either city or mine for Limo and Nissian respectively
Completion of entire game including bonus stages for the 6x6